Move integration cases into tests package
