+++ raftstore/worker_pool_test.go
@@ -25,6 +25,7 @@ import (
 func TestWorkerPoolCanBeCreatedAndClosed(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	p := newWorkerPool(nil, nil, 32)
+	p.start()
 	p.close()
 }
 
@@ -81,6 +82,7 @@ func TestWorkerPoolCanScheduleSimpleJob(t *testing.T) {
 	l := newTestReplicaLoader()
 	h, _ := l.getReplica(10)
 	p := newWorkerPool(nil, l, 32)
+	p.start()
 	defer func() {
 		p.close()
 		assert.Equal(t, 0, len(p.pending))
@@ -103,6 +105,7 @@ func TestWorkerPoolCanScheduleSimpleJob(t *testing.T) {
 func TestWorkerPoolWillNotReturnBusyWorker(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	p := newWorkerPool(nil, nil, 32)
+	p.start()
 	defer p.close()
 	assert.Equal(t, 32, len(p.workers))
 	assert.Equal(t, 0, len(p.busy))
@@ -119,6 +122,7 @@ func TestWorkerPoolWillNotReturnBusyWorker(t *testing.T) {
 func TestWorkerPoolScheduleNothingWhenNotPendingJob(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	p := newWorkerPool(nil, nil, 32)
+	p.start()
 	defer p.close()
 	assert.False(t, p.scheduleWorker())
 	w := p.getWorker()
@@ -128,6 +132,7 @@ func TestWorkerPoolScheduleNothingWhenNotPendingJob(t *testing.T) {
 func TestWorkerPoolScheduleNothingWhenNoIdleWorker(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	p := newWorkerPool(nil, nil, 32)
+	p.start()
 	defer p.close()
 	p.pending[20] = nil
 	for _, w := range p.workers {
@@ -141,6 +146,7 @@ func TestWorkerPoolScheduleNothingWhenNoIdleWorker(t *testing.T) {
 func TestWorkerPoolWillNotConcurrentlyProcessTheSameShard(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	p := newWorkerPool(nil, nil, 32)
+	p.start()
 	defer p.close()
 	p.pending[10] = nil
 	p.processing[10] = struct{}{}
@@ -151,6 +157,7 @@ func TestWorkerPoolSetBusyAndProcessingAsExpected(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	l := newTestReplicaLoader()
 	p := newWorkerPool(nil, l, 32)
+	p.start()
 	defer func() {
 		p.close()
 		assert.Equal(t, 0, len(p.busy))
@@ -169,6 +176,7 @@ func testWorkerPoolConcurrentJobs(t *testing.T, moreJob bool) {
 	defer leaktest.AfterTest(t)()
 	l := newTestReplicaLoader()
 	p := newWorkerPool(nil, l, 32)
+	p.start()
 	defer func() {
 		p.close()
 		assert.Equal(t, 0, len(p.busy))
