+++ test/doc.go
@@ -1,0 +1,8 @@
+/*
+The test package is used to store all integration tests. Integration tests start multiple `Stores` and `Replica`
+in a process to test whether the `Cube` is working properly from a usage point of view.
+
+
+*/
+
+package test
